Home banner slide build
       - The Home storyboard view controller wasn't connected, so the storyboard was loaded to use the Home view controller properly.
       - Using compositional layout and diffable data source to efficiently manage and display collection view items.
