+++ Store-ios/Feature/Home/HomeBannerCollectionViewCell.swift
@@ -1,0 +1,21 @@
+//
+//  HomeBannerCollectionViewCell.swift
+//  Store-ios
+//
+//  Created by TaeWook Park on 9/1/24.
+//
+
+import UIKit
+
+// Banner slide
+class HomeBannerCollectionViewCell: UICollectionViewCell {
+    
+    // slide image view 
+    @IBOutlet weak var imageView: UIImageView!
+    
+    
+    func setImage(_ image: UIImage) {
+        imageView.image = image 
+    }
+    
+}
