--- Store-ios/HomeViewController.swift
@@ -1,19 +1,0 @@
-//
-//  HomeViewController.swift
-//  Store-ios
-//
-//  Created by TaeWook Park on 9/1/24.
-//
-
-import UIKit
-
-class HomeViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        
-        view.backgroundColor = .uclaLighterBlue
-
-    }
-
-}
